Support multiple JS examples
